+++ hero_ship.h
@@ -5,7 +5,7 @@
 class Hero_Ship : public Floating_Object
 {
 public:
-    Hero_Ship(float t_X, float t_Y, const sf::Texture *texture);
+    Hero_Ship(int t_X, int t_Y, const sf::Texture *texture);
     ~Hero_Ship() = default;
     Hero_Ship() = delete;
     void hero_update();  // function used for updating state of object
